requirements.txt
backcall==0.2.0
bleach==5.0.1
blosc==1.10.6
boto3==1.24.81
botocore==1.27.81
catboost==1.1
charset-normalizer==2.1.1
colorama==0.4.5
commonmark==0.9.1
cycler==0.11.0
dataclasses==0.6
debugpy==1.6.3
decorator==5.1.1
dill==0.3.5.1
dll-diagnostics==0.0.18
dnspython==2.2.1
docutils==0.19
entrypoints==0.4
evidently==0.1.57.dev0
fonttools==4.37.3
from-root==1.0.2
future==0.18.2
graphviz==0.20.1
idna==3.4
imbalanced-learn==0.9.0
importlib-metadata==4.12.0
ipykernel==6.16.0
ipython==7.34.0
jaraco.classes==3.2.3
jedi==0.18.1
jmespath==1.0.1
joblib==1.2.0
jupyter-core==4.11.1
jupyter_client==7.3.5
keyring==23.9.3
kiwisolver==1.4.4
kneed==0.8.1
lightgbm==3.3.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
miceforest==5.6.2
more-itertools==8.14.0
nest-asyncio==1.5.5
networkx==2.6.3
numpy==1.21.6
packaging==21.3
pandas==1.3.5
parso==0.8.3
patsy==0.5.2
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.2.0
pkginfo==1.8.3
plotly==5.10.0
prettytable==3.4.1
prompt-toolkit==3.0.31
psutil==5.9.2
pydot==1.4.2
Pygments==2.13.0
pymongo==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==5.4.1
pyzmq==24.0.1
readme-renderer==37.2
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
s3transfer==0.6.0
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.0
six==1.16.0
statsmodels==0.13.2
tenacity==8.1.0
termcolor==2.0.1
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.4.0
twine==4.0.1
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
wincertstore==0.2
xgboost==1.6.2
zipp==3.8.1